\name{specs-verification-package}
\title{Verification routines for the SPECS FP7 project}
\description{
A collection of forecast verification routines, with emphasis on seasonal-to-decadal climate ensemble forecasts.
}
\usage{
see individual functions for usage
}
\arguments{
see individual functions for arguments
}
\examples{
## examples
}
